+++ src/main/webapp/WEB-INF/views/home.jsp
@@ -18,19 +18,23 @@
 					</h3>
 					<div class="dropdown-content">
 						<a href="viewConsultant/${user.consultantId}"
-							class="btn btn-outline-success"> View/Update Profile </a> <a
-							href="logout" class="btn btn-outline-success"> Log Out </a>
+							class="btn btn-primary"> View/Update Profile </a> <a
+							href="logout" class="btn btn-primary"> Log Out </a>
 					</div>
 				</div>
 				<form action="homePage" method="get">
 					<table class="table table-light">
 					  <tr>
-						<td><input type="text" id="searchName" name="searchName" value="${searchName}" placeholder="enter name" style="width: 200px; "/>
-						<input type="text" id="searchMobile" name="searchMobile" value="${searchMobile}" placeholder="enter mobilenumber" style="width: 180px; "/>
-						<input type="text" id="searchEmail" name="searchEmail" value="${searchEmail}" placeholder="enter email" style="width: 200px; "/>
-						<input type="text" id="searchAge" name="searchAge" value="${searchAge}" placeholder="enter age" style="width: 180px; "/>
-						<input type="text" id="searchDiagnosis" name="searchDiagnosis" value="${searchDiagnosis}" placeholder="enter diagnosis" style="width: 200px; "/>
-						<button type="submit" class="btn btn-primary">Search</button></td>
+						<td>
+							<input type="text" id="searchName" name="searchName" value="${searchName}" placeholder="enter name" style="width: 180px; "/>
+							<input type="text" id="searchMobile" name="searchMobile" value="${searchMobile}" placeholder="enter mobilenumber" style="width: 180px; "/>
+							<input type="text" id="searchEmail" name="searchEmail" value="${searchEmail}" placeholder="enter email" style="width: 180px; "/>
+							<input type="text" id="searchPlace" name="searchPlace" value="${searchPlace}" placeholder="enter place" style="width: 180px; "/>
+							<input type="text" id="searchDiagnosis" name="searchDiagnosis" value="${searchDiagnosis}" placeholder="enter diagnosis" style="width: 180px; "/>
+							<button type="submit" class="btn btn-primary">Search</button>
+							<button type="button" class="btn btn-primary" onclick="window.location.href='homePage?searchName=&searchMobile=&searchEmail=&searchPlace=&searchDiagnosis=&pageSize=5'">
+							 Reset</button>
+						</td>
 					  </tr>
 					</table>
 					<table>
@@ -73,11 +77,11 @@
 				<table class="table table-light">
 					<thead class="thead-dark">
 						<tr>
-							<th scope="col">Patient Number</th>
-							<th scope="col">Patient Name</th>
-							<th scope="col">Mobile Number</th>
-							<th scope="col">Email</th>
-							<th scope="col">Place</th>
+							<th scope="col" style="width: 100px;">Patient Number</th>
+							<th scope="col" style="width: 180px;">Patient Name</th>
+							<th scope="col" style="width: 100px;">Mobile Number</th>
+							<th scope="col" style="width: 100px;">Email</th>
+							<th scope="col" style="width: 100px;">Place</th>
 						</tr>
 					</thead>
 					<tbody>
@@ -93,18 +97,18 @@
 										<i class="fa fa-eye"></i> </a> 
 									<a href="updatePatient/${p.id}" title="Update Patient"><i
 										class="fa-sharp fa-solid fa-pen-to-square text-success"></i></a>
-									<a href="deletePatient/${p.id}" title="Delete Patient"><i
+									<a href="deletePatient/${p.id}" title="Delete Patient" onclick="return confirm('Are you sure you want to delete record?');"><i
 										class="fa-sharp fa-solid fa-trash text-danger"></i></a>
-								    <a href="addSessionNotes/${p.id}" class="btn btn-outline-success">
+								    <a href="addSessionNotes/${p.id}" class="btn btn-primary">
 										Sessions </a>
-									<a href="add-document-${p.id}" class="btn btn-outline-success"> Attachments </a>
+									<a href="add-document-${p.id}" class="btn btn-primary"> Attachments </a>
 								</td>
 							</tr>
 						</c:forEach>
 					</tbody>
 				</table>
 				<div class="container text-center mb-3">
-					<a href="addPatient" class="btn btn-outline-success"> Add
+					<a href="addPatient" class="btn btn-primary"> Add
 						Patient </a>
 				</div>
